+++ src/Clusterers/KMeans.php
@@ -11,7 +11,9 @@ use Rubix\ML\Other\Helpers\Params;
 use Rubix\ML\Other\Helpers\DataType;
 use Rubix\ML\Other\Traits\LoggerAware;
 use Rubix\ML\Kernels\Distance\Distance;
+use Rubix\ML\Clusterers\Seeders\Seeder;
 use Rubix\ML\Kernels\Distance\Euclidean;
+use Rubix\ML\Clusterers\Seeders\PlusPlus;
 use Rubix\ML\Other\Specifications\DatasetIsCompatibleWithEstimator;
 use InvalidArgumentException;
 use RuntimeException;
@@ -25,8 +27,7 @@ use RuntimeException;
  * within cluster distance as a loss function.
  *
  * References:
- * [1] D. Arthur et al. (2006). k-means++: The Advantages of Careful Seeding.
- * [2] D. Sculley. (2010). Web-Scale K-Means Clustering.
+ * [1] D. Sculley. (2010). Web-Scale K-Means Clustering.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -58,6 +59,14 @@ class KMeans implements Learner, Persistable, Verbose
     protected $kernel;
 
     /**
+     * The maximum number of iterations to run until the algorithm
+     * terminates.
+     *
+     * @var int
+     */
+    protected $epochs;
+
+    /**
      * The minimum change in the size of each cluster for training
      * to continue.
      *
@@ -66,12 +75,11 @@ class KMeans implements Learner, Persistable, Verbose
     protected $minChange;
 
     /**
-     * The maximum number of iterations to run until the algorithm
-     * terminates.
+     * The cluster centroid seeder.
      *
-     * @var int
+     * @var \Rubix\ML\Clusterers\Seeders\Seeder
      */
-    protected $epochs;
+    protected $seeder;
 
     /**
      * The computed centroid vectors of the training data.
@@ -86,16 +94,18 @@ class KMeans implements Learner, Persistable, Verbose
      * @param int $k
      * @param int $batchSize
      * @param \Rubix\ML\Kernels\Distance\Distance|null $kernel
-     * @param int $minChange
      * @param int $epochs
+     * @param int $minChange
+     * @param \Rubix\ML\Clusterers\Seeders\Seeder $seeder
      * @throws \InvalidArgumentException
      */
     public function __construct(
         int $k,
         int $batchSize = 100,
         ?Distance $kernel = null,
+        int $epochs = 300,
         int $minChange = 1,
-        int $epochs = 300
+        ?Seeder $seeder = null
     ) {
         if ($k < 1) {
             throw new InvalidArgumentException('Must target at least'
@@ -107,21 +117,22 @@ class KMeans implements Learner, Persistable, Verbose
                 . " than 0, $batchSize given.");
         }
 
-        if ($minChange < 1) {
-            throw new InvalidArgumentException('Min change cannot be less'
-                . " than 1, $minChange given.");
-        }
-
         if ($epochs < 1) {
             throw new InvalidArgumentException('Estimator must train'
                 . " for at least 1 epoch, $epochs given.");
         }
 
+        if ($minChange < 1) {
+            throw new InvalidArgumentException('Min change cannot be less'
+                . " than 1, $minChange given.");
+        }
+
         $this->k = $k;
         $this->batchSize = $batchSize;
         $this->kernel = $kernel ?? new Euclidean();
-        $this->minChange = $minChange;
         $this->epochs = $epochs;
+        $this->minChange = $minChange;
+        $this->seeder = $seeder ?? new PlusPlus($kernel);
     }
 
     /**
@@ -180,12 +191,15 @@ class KMeans implements Learner, Persistable, Verbose
             $this->logger->info('Learner initialized w/ '
                 . Params::stringify([
                     'k' => $this->k,
+                    'batch_size' => $this->batchSize,
                     'kernel' => $this->kernel,
                     'epochs' => $this->epochs,
+                    'min_change' => $this->minChange,
+                    'seeder' => $this->seeder,
                 ]));
         }
 
-        $this->centroids = $this->initialize($dataset);
+        $this->centroids = $this->seeder->seed($dataset, $this->k);
 
         $samples = $dataset->samples();
         $labels = array_fill(0, $dataset->numRows(), null);
@@ -274,7 +288,6 @@ class KMeans implements Learner, Persistable, Verbose
      * Label a given sample based on its distance from a particular centroid.
      *
      * @param array $sample
-     * @throws \RuntimeException
      * @return int
      */
     protected function assign(array $sample) : int
@@ -292,58 +305,5 @@ class KMeans implements Learner, Persistable, Verbose
         }
 
         return (int) $bestCluster;
-    }
-
-    /**
-     * Initialize the cluster centroids using the k-means++ method.
-     *
-     * @param \Rubix\ML\Datasets\Dataset $dataset
-     * @throws \RuntimeException
-     * @return array
-     */
-    protected function initialize(Dataset $dataset) : array
-    {
-        $n = $dataset->numRows();
-
-        if ($n < $this->k) {
-            throw new RuntimeException('The number of samples cannot be less'
-                . ' than the number of target clusters.');
-        }
-
-        $weights = array_fill(0, $n, 1. / $n);
-
-        $centroids = [];
-
-        for ($i = 0; $i < $this->k; $i++) {
-            $subset = $dataset->randomWeightedSubsetWithReplacement(1, $weights);
-
-            $centroids[] = $subset[0];
-
-            if ($i === $this->k) {
-                break 1;
-            }
-
-            foreach ($dataset as $j => $sample) {
-                $closest = INF;
-
-                foreach ($centroids as $centroid) {
-                    $distance = $this->kernel->compute($sample, $centroid);
-
-                    if ($distance < $closest) {
-                        $closest = $distance;
-                    }
-                }
-
-                $weights[$j] = $closest ** 2;
-            }
-
-            $total = array_sum($weights) ?: self::EPSILON;
-
-            foreach ($weights as &$weight) {
-                $weight /= $total;
-            }
-        }
-
-        return $centroids;
     }
 }