+++ src/Clusterers/Seeders/PlusPlus.php
@@ -1,0 +1,96 @@
+<?php
+
+namespace Rubix\ML\Clusterers\Seeders;
+
+use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Kernels\Distance\Distance;
+use Rubix\ML\Kernels\Distance\Euclidean;
+use RuntimeException;
+
+/**
+ * Plus Plus
+ *
+ * This seeder attempts to maximize the likelihood of seeding distant clusters
+ * while still remaining random. It does so by sequentially selecting random
+ * samples weighted by their distance from the previous seed.
+ *
+ * References:
+ * [1] D. Arthur et al. (2006). k-means++: The Advantages of PlusPlus Seeding.
+ * [2] A. Stetco et al. (2015). Fuzzy C-means++: Fuzzy C-means with effective
+ * seeding initialization.
+ *
+ * @category    Machine Learning
+ * @package     Rubix/ML
+ * @author      Andrew DalPino
+ */
+class PlusPlus implements Seeder
+{
+    /**
+     * The distance kernel used to compute the distance between samples.
+     *
+     * @var \Rubix\ML\Kernels\Distance\Distance
+     */
+    protected $kernel;
+
+    /**
+     * @param \Rubix\ML\Kernels\Distance\Distance|null $kernel
+     */
+    public function __construct(?Distance $kernel = null)
+    {
+        $this->kernel = $kernel ?? new Euclidean();
+    }
+
+    /**
+     * Seed k cluster centroids from a dataset.
+     *
+     * @param \Rubix\ML\Datasets\Dataset $dataset
+     * @param int $k
+     * @throws \RuntimeException
+     * @return array
+     */
+    public function seed(Dataset $dataset, int $k) : array
+    {
+        $n = $dataset->numRows();
+
+        if ($n < $k) {
+            throw new RuntimeException('The number of samples cannot be less'
+                . ' than the number of target clusters.');
+        }
+
+        $weights = array_fill(0, $n, 1. / $n);
+
+        $centroids = [];
+
+        for ($i = 0; $i < $k; $i++) {
+            $subset = $dataset->randomWeightedSubsetWithReplacement(1, $weights);
+
+            $centroids[] = $subset[0];
+
+            if ($i === $k) {
+                break 1;
+            }
+
+            foreach ($dataset as $j => $sample) {
+                $closest = INF;
+
+                foreach ($centroids as $centroid) {
+                    $distance = $this->kernel->compute($sample, $centroid);
+
+                    if ($distance < $closest) {
+                        $closest = $distance;
+                    }
+                }
+
+                $weights[$j] = $closest ** 2;
+            }
+
+            $total = array_sum($weights) ?: self::EPSILON;
+
+            foreach ($weights as &$weight) {
+                $weight /= $total;
+            }
+        }
+
+        return $centroids;
+    }
+}