+++ src/Clusterers/Seeders/Random.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Rubix\ML\Clusterers\Seeders;
+
+use Rubix\ML\Datasets\Dataset;
+
+/**
+ * Random
+ *
+ * Completely random selection of k seeds from the given dataset.
+ *
+ * @category    Machine Learning
+ * @package     Rubix/ML
+ * @author      Andrew DalPino
+ */
+class Random implements Seeder
+{
+    /**
+     * Seed k cluster centroids from a dataset.
+     *
+     * @param \Rubix\ML\Datasets\Dataset $dataset
+     * @param int $k
+     * @return array
+     */
+    public function seed(Dataset $dataset, int $k) : array
+    {
+        return $dataset->randomize()->head($k)->samples();
+    }
+}