+++ tests/Clusterers/KMeansTest.php
@@ -8,10 +8,11 @@ use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Clusterers\KMeans;
 use Rubix\ML\Datasets\Unlabeled;
-use Rubix\ML\Other\Loggers\BlackHole;
 use Rubix\ML\Other\Helpers\DataType;
+use Rubix\ML\Other\Loggers\BlackHole;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Kernels\Distance\Euclidean;
+use Rubix\ML\Clusterers\Seeders\PlusPlus;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 use Rubix\ML\CrossValidation\Metrics\VMeasure;
 use PHPUnit\Framework\TestCase;
@@ -40,7 +41,7 @@ class KMeansTest extends TestCase
             'blue' => new Blob([0, 0, 255], 20.),
         ]);
 
-        $this->estimator = new KMeans(3, 100, new Euclidean(), 1, 300);
+        $this->estimator = new KMeans(3, 100, new Euclidean(), 100, 1, new PlusPlus(new Euclidean()));
 
         $this->estimator->setLogger(new BlackHole());